+++ src/param.cc
@@ -108,8 +108,8 @@ void Param::parse(Model &model) {
     argv_i = argv_[argc_i];
 
     if (argc_i == 0) {
-      sample_size = readNextInput<size_t>();
-      model.set_loci_number(readNextInput<size_t>());
+      sample_size = readNextInt();
+      model.set_loci_number(readNextInt());
     }
 
     // ------------------------------------------------------------------
@@ -131,7 +131,7 @@ void Param::parse(Model &model) {
     // ------------------------------------------------------------------
     else if (argv_i == "-r") {
       par_bool = readNextInput<double>();
-      model.setLocusLength(readNextInput<size_t>());
+      model.setLocusLength(readNextInt());
       model.setRecombinationRate(par_bool, true, true, 0.0);
     }
 
@@ -145,10 +145,10 @@ void Param::parse(Model &model) {
     // ------------------------------------------------------------------
     // Set number of subpopulations and samples at time 0
     else if (argv_i == "-I") {
-      model.set_population_number(readNextInput<size_t>());
+      model.set_population_number(readNextInt());
       std::vector<size_t> sample_size;
       for (size_t i = 0; i < model.population_number(); ++i) {
-        sample_size.push_back(readNextInput<size_t>());
+        sample_size.push_back(readNextInt());
       }
       model.addSampleSizes(0.0, sample_size);
       // there might or might not follow a symmetric migration rate
@@ -168,7 +168,7 @@ void Param::parse(Model &model) {
       }
       std::vector<size_t> sample_size;
       for (size_t i = 0; i < model.population_number(); ++i) {
-        sample_size.push_back(readNextInput<size_t>());
+        sample_size.push_back(readNextInt());
       }
       model.addSampleSizes(time, sample_size, true);
     }
@@ -190,7 +190,7 @@ void Param::parse(Model &model) {
     else if (argv_i == "-en" || argv_i == "-n") {
       if (argv_i == "-en") time = readNextInput<double>();
       else time = 0.0;
-      size_t pop = readNextInput<size_t>() - 1;
+      size_t pop = readNextInt() - 1;
       model.addPopulationSize(time, pop, readNextInput<double>(), true, true);
       if (time != 0.0) model.addGrowthRate(time, pop, 0.0, true);
     }
@@ -211,7 +211,7 @@ void Param::parse(Model &model) {
     else if (argv_i == "-g" || argv_i == "-eg") {
       if (argv_i == "-eg") time = readNextInput<double>();
       else time = 0.0;
-      size_t pop = readNextInput<size_t>() - 1;
+      size_t pop = readNextInt() - 1;
       model.addGrowthRate(time, pop, readNextInput<double>(), true, true); 
     }
 
@@ -249,8 +249,8 @@ void Param::parse(Model &model) {
         throw std::invalid_argument(std::string("If you use '-m' or '-em' in a model with population merges ('-es'),") +
                                     std::string("then you need to sort both arguments by time."));
       }
-      source_pop = readNextInput<size_t>() - 1;
-      sink_pop = readNextInput<size_t>() - 1;
+      source_pop = readNextInt() - 1;
+      sink_pop = readNextInt() - 1;
 
       model.addMigrationRate(time, source_pop, sink_pop, readNextInput<double>(), true, true);
     }
@@ -277,7 +277,7 @@ void Param::parse(Model &model) {
                                     std::string("by the time they occur."));
       }
       min_time = time;
-      source_pop = readNextInput<size_t>() - 1;
+      source_pop = readNextInt() - 1;
       sink_pop = model.population_number();
       double fraction = readNextInput<double>();
 
@@ -291,8 +291,8 @@ void Param::parse(Model &model) {
     // ------------------------------------------------------------------
     else if (argv_i == "-ej") {
       time = readNextInput<double>();
-      source_pop = readNextInput<size_t>() - 1;
-      sink_pop = readNextInput<size_t>() - 1;
+      source_pop = readNextInt() - 1;
+      sink_pop = readNextInt() - 1;
       model.addSingleMigrationEvent(time, source_pop, sink_pop, 1.0, true); 
       for (size_t i = 0; i < model.population_number(); ++i) {
         if (i == source_pop) continue;
@@ -304,7 +304,7 @@ void Param::parse(Model &model) {
     // Pruning 
     // ------------------------------------------------------------------
     else if (argv_i == "-l"){
-      model.set_exact_window_length(readNextInput<size_t>());
+      model.set_exact_window_length(readNextInt());
     }
 
     // ------------------------------------------------------------------
@@ -358,7 +358,7 @@ void Param::parse(Model &model) {
     }
 
     else if (argv_i == "-p") {
-      this->precision_ = readNextInput<size_t>() ;
+      this->precision_ = readNextInt() ;
     }
 
     // ------------------------------------------------------------------
@@ -367,10 +367,10 @@ void Param::parse(Model &model) {
     else if (argv_i == "-seed" || argv_i == "--seed") {
       std::vector<size_t> seeds(3, 0);
       // Always require one seed
-      seeds.at(0) = readNextInput<size_t>();
+      seeds.at(0) = readNextInt();
       try {
         // Maybe read in up to 3 seeds (ms compatibility)
-        for (size_t i = 1; i < 3; i++) seeds.at(i) = readNextInput<size_t>();
+        for (size_t i = 1; i < 3; i++) seeds.at(i) = readNextInt();
       } catch (std::invalid_argument e) {
         --argc_i;
       }